server/requirements.txt
flask>=2.3.0
ollama>=0.4.7
llama-cpp-python